Rename common_request_handler to concurrent_common_request_handler.
