+++ opentracing-testbed/src/test/java/io/opentracing/testbed/concurrent_common_request_handler/HandlerTest.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.opentracing.testbed.common_request_handler;
+package io.opentracing.testbed.concurrent_common_request_handler;
 
 import io.opentracing.Scope;
 import io.opentracing.Span;
